Remove unused code
This code seems to be unnecessary.  It is operating on `@s` after it has
already been used to create `@/`.  The value in `@s` is changed right
before it is overwritten.

Thinking that this may be a bug I tried changing `let @s = '\V' . @s` to
`let @/ = '\V' . @/`, but this broke visual `<leader>*` (specifically on the
`**` line 12 in the test-patterns).  It seems this code is obsolete.

Conflicts:
	plugin/visual-star-search.vim